feat: dialog box

+++ src/app/(task)/all-tasks/AllTasks.tsx
@@ -6,6 +6,7 @@ import { useContext, useEffect, useState } from 'react';
 import { DataTable } from './data-table';
 import { columns } from './columns';
 import { toast } from 'react-toastify';
+import Swal from 'sweetalert2';
 
 export type TypeTasksTable = {
     content: string;
@@ -51,12 +52,24 @@ export default AllTasks;
 
 export async function deleteTask(taskId: string) {
     try {
-        const result = await deleteUserTask(taskId);
-        if (result.success === false) {
-            toast.error(result.message);
-            return;
-        }
-        toast.success(result.message);
+        Swal.fire({
+            title: 'Are you sure?',
+            text: "You won't be able to revert this!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Yes, delete it!',
+        }).then(async (result) => {
+            if (result.isConfirmed) {
+                const result = await deleteUserTask(taskId);
+                if (result.success === false) {
+                    toast.error(result.message);
+                    return;
+                }
+                toast.success(result.message);
+            }
+        });
     } catch (error: any) {
         toast.error('Something went wrong!');
     }
